+++ src/main/java/pl/edu/wszib/what/todo/notes/dao/impl/IUserDAO.java
@@ -6,12 +6,12 @@ import java.util.List;
 import java.util.Optional;
 
 public interface IUserDAO {
-    Optional<User> getById(int id);
+    Optional<User> getById(Long id);
     Optional<User> getByLogin(String login);
 
     List<User> getAll();
 
     void save(User user);
-    void delete(int id);
+    void delete(Long id);
     void update(User user);
 }
